export functionality for building runtime with scuttling config
