+++ packages/lavapack/src/index.js
@@ -1,2 +1,3 @@
 module.exports = require('./pack.js')
 module.exports.makePolicyLoaderStream = require('./makePolicyLoaderStream.js')
+module.exports.buildRuntime = require('./builder-runtime.js')
